requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
asyncpg==0.29.0

# Binance API
python-binance==1.0.19
ccxt==4.1.68

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7
pyjwt==2.8.0

# Machine Learning & Data Science
tensorflow==2.15.0
torch==2.1.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Technical Analysis
TA-Lib==0.4.28
mplfinance==0.12.10a0

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0

# Async Operations
aiohttp==3.9.1
asyncio==3.4.3
aioredis==2.0.1

# Monitoring & Logging
python-json-logger==2.0.7
loguru==0.7.2

# Notifications
python-telegram-bot==20.7
smtplib-async==1.0.0

# Environment & Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0